code/requirements.txt
pandas>=1.5.0
numpy>=1.21.0
python-dotenv>=0.19.0
tqdm>=4.62.0
scikit-learn>=0.24.0
requests>=2.25.0 
kiwi>=1.4.0